requirements.txt
scapy>=2.5.0
PyYAML>=6.0.1
cryptography>=42.0.0
pyOpenSSL>=24.0.0
watchdog>=4.0.0
